Move the generate type object command to a separate file so it doesn't run with tests
