+++ modules/product-catalog/src/generateTypeObject.ts
@@ -1,7 +1,5 @@
-import fs from 'fs';
 import { arrayToObject, distinct } from '@modules/arrayFunctions';
 import { checkDefined } from '@modules/nullAndUndefined';
-import { getZuoraCatalogFromS3 } from '@modules/zuora-catalog/S3';
 import type {
 	ZuoraCatalog,
 	ZuoraProductRatePlan,
@@ -77,18 +75,3 @@ export const generateTypeObject = (catalog: ZuoraCatalog) => {
 
 	return arrayToObject(arrayVersion);
 };
-
-const writeTypesToFile = async () => {
-	const prodCatalog = await getZuoraCatalogFromS3('PROD');
-	const types = generateTypeObject(prodCatalog);
-	const typesString = JSON.stringify(types, null, 2);
-
-	fs.writeFileSync(
-		'./src/typeObject.ts',
-		`export const typeObject = ${typesString} as const;`,
-	);
-};
-
-void (async function () {
-	await writeTypesToFile();
-})();
